requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==3.14.0
gunicorn==20.1.0
python-dotenv==1.0.0